Adds partial refactor of users create endpoint
